Serialization of injector works!!!!!
