+++ projects/interactions/private/pybindings/DarkNewsDecay.h
@@ -7,272 +7,8 @@
 #include <pybind11/stl.h>
 #include <pybind11/embed.h>
 
+#include "../../public/SIREN/interactions/Decay.h"
 #include "../../public/SIREN/interactions/DarkNewsDecay.h"
-#include "../../../dataclasses/public/SIREN/dataclasses/Particle.h"
-#include "../../../dataclasses/public/SIREN/dataclasses/InteractionRecord.h"
-#include "../../../dataclasses/public/SIREN/dataclasses/InteractionSignature.h"
-#include "../../../geometry/public/SIREN/geometry/Geometry.h"
-#include "../../../utilities/public/SIREN/utilities/Random.h"
-#include "../../../utilities/public/SIREN/utilities/Pybind11Trampoline.h"
-
-namespace siren {
-namespace interactions {
-// Trampoline class for Decay
-class pyDecay : public Decay,Pybind11Trampoline<Decay, pyDecay> {
-public:
-    using Decay::Decay;
-    pyDecay(Decay && parent) : Decay(std::move(parent)) {}
-    pybind11::object self;
-
-    double TotalDecayLength(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE(
-            self,
-            Decay,
-            double,
-            TotalDecayLength,
-            "TotalDecayLength",
-            interaction
-        )
-    }
-
-    double TotalDecayLengthForFinalState(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE(
-            self,
-            Decay,
-            double,
-            TotalDecayLengthForFinalState,
-            "TotalDecayLengthForFinalState",
-            interaction
-        )
-    }
-    
-    double TotalDecayWidth(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            Decay,
-            double,
-            TotalDecayWidth,
-            "TotalDecayWidth",
-            interaction
-        )
-    }
-
-    double TotalDecayWidthForFinalState(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            Decay,
-            double,
-            TotalDecayWidthForFinalState,
-            "TotalDecayWidthForFinalState",
-            interaction
-        )
-    }
-
-    double TotalDecayWidth(siren::dataclasses::ParticleType primary) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            Decay,
-            double,
-            TotalDecayWidth,
-            "TotalDecayWidth",
-            primary
-        )
-    }
-
-    double DifferentialDecayWidth(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            Decay,
-            double,
-            DifferentialDecayWidth,
-            "DifferentialDecayWidth",
-            interaction
-        )
-    }
-
-    void SampleFinalState(dataclasses::CrossSectionDistributionRecord & record, std::shared_ptr<siren::utilities::SIREN_random> random) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            Decay,
-            void,
-            SampleFinalState,
-            "SampleFinalState",
-            record,
-            random
-        )
-    }
-
-    std::vector<siren::dataclasses::InteractionSignature> GetPossibleSignatures() const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            Decay,
-            std::vector<siren::dataclasses::InteractionSignature>,
-            GetPossibleSignatures,
-            "GetPossibleSignatures"
-        )
-    }
-
-    std::vector<siren::dataclasses::InteractionSignature> GetPossibleSignaturesFromParent(siren::dataclasses::ParticleType primary_type) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            Decay,
-            std::vector<siren::dataclasses::InteractionSignature>,
-            GetPossibleSignaturesFromParents,
-            "GetPossibleSignaturesFromParents",
-            primary_type
-        )
-    }
-
-    std::vector<std::string> DensityVariables() const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            Decay,
-            std::vector<std::string>,
-            DensityVariables,
-            "DensityVariables"
-        )
-    }
-
-    double FinalStateProbability(dataclasses::InteractionRecord const & record) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            Decay,
-            double,
-            FinalStateProbability,
-            "FinalStateProbability",
-            record
-        )
-    }
-
-};
-// Trampoline class for DarkNewsDecay
-class pyDarkNewsDecay : public DarkNewsDecay,public Pybind11Trampoline<DarkNewsDecay, pyDarkNewsDecay> {
-public:
-    using DarkNewsDecay::DarkNewsDecay;
-    pyDarkNewsDecay(DarkNewsDecay && parent) : DarkNewsDecay(std::move(parent)) {
-        self = pybind11::reinterpret_borrow<pybind11::object>(pybind11::handle(get_object_handle(&parent, pybind11::detail::get_type_info(typeid(DarkNewsDecay)))));
-    }
-    pyDarkNewsDecay(DarkNewsDecay const & parent) : DarkNewsDecay(parent) {
-        self = pybind11::reinterpret_borrow<pybind11::object>(pybind11::handle(get_object_handle(&parent, pybind11::detail::get_type_info(typeid(DarkNewsDecay)))));
-    }
-    pybind11::object self;
-
-    double TotalDecayWidth(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsDecay,
-            double,
-            TotalDecayWidth,
-            "TotalDecayWidth",
-            std::cref(interaction)
-        )
-    }
-
-    double TotalDecayWidthForFinalState(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsDecay,
-            double,
-            TotalDecayWidthForFinalState,
-            "TotalDecayWidthForFinalState",
-            std::cref(interaction)
-        )
-    }
-
-    double TotalDecayWidth(siren::dataclasses::ParticleType primary) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsDecay,
-            double,
-            TotalDecayWidth,
-            "TotalDecayWidth",
-            primary
-        )
-    }
-
-    double DifferentialDecayWidth(dataclasses::InteractionRecord const & interaction) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsDecay,
-            double,
-            DifferentialDecayWidth,
-            "DifferentialDecayWidth",
-            std::cref(interaction)
-        )
-    }
-
-    void SampleRecordFromDarkNews(dataclasses::CrossSectionDistributionRecord & record, std::shared_ptr<siren::utilities::SIREN_random> random) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsDecay,
-            void,
-            SampleRecordFromDarkNews,
-            "SampleRecordFromDarkNews",
-            std::ref(record),
-            random
-        )
-    }
-
-    void SampleFinalState(dataclasses::CrossSectionDistributionRecord & record, std::shared_ptr<siren::utilities::SIREN_random> random) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsDecay,
-            void,
-            SampleFinalState,
-            "SampleFinalState",
-            std::ref(record),
-            random
-        )
-    }
-
-    std::vector<siren::dataclasses::InteractionSignature> GetPossibleSignatures() const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            DarkNewsDecay,
-            std::vector<siren::dataclasses::InteractionSignature>,
-            GetPossibleSignatures,
-            "GetPossibleSignatures"
-        )
-    }
-
-    std::vector<siren::dataclasses::InteractionSignature> GetPossibleSignaturesFromParent(siren::dataclasses::ParticleType primary_type) const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            DarkNewsDecay,
-            std::vector<siren::dataclasses::InteractionSignature>,
-            GetPossibleSignaturesFromParent,
-            "GetPossibleSignaturesFromParent",
-            primary_type
-        )
-    }
-
-    std::vector<std::string> DensityVariables() const override {
-        SELF_OVERRIDE_PURE(
-            self,
-            DarkNewsDecay,
-            std::vector<std::string>,
-            DensityVariables,
-            "DensityVariables"
-        )
-    }
-
-    double FinalStateProbability(dataclasses::InteractionRecord const & record) const override {
-        SELF_OVERRIDE(
-            self,
-            DarkNewsDecay,
-            double,
-            FinalStateProbability,
-            "FinalStateProbability",
-            std::cref(record)
-        )
-    }
-    
-    pybind11::object get_representation() {
-        return self;
-    }
-};
-} // end interactions namespace
-} // end LI namespace
 
 void register_DarkNewsDecay(pybind11::module_ & m) {
     using namespace pybind11;
