+++ projects/interactions/public/SIREN/interactions/DarkNewsDecay.h
@@ -22,6 +22,7 @@
 
 #include "SIREN/interactions/Decay.h"   // for Decay
 #include "SIREN/dataclasses/Particle.h"  // for Particle
+#include "SIREN/utilities/Pybind11Trampoline.h" // for Pybind11Trampoline
 
 namespace siren { namespace dataclasses { class InteractionRecord; } }
 namespace siren { namespace dataclasses { class CrossSectionDistributionRecord; } }
@@ -69,7 +70,140 @@ public:
         }
     }
 
-};
+}; // class DarkNewsDecay
+
+// Trampoline class for DarkNewsDecay
+class pyDarkNewsDecay : public DarkNewsDecay,public Pybind11Trampoline<DarkNewsDecay, pyDarkNewsDecay> {
+public:
+    using DarkNewsDecay::DarkNewsDecay;
+    pyDarkNewsDecay(DarkNewsDecay && parent) : DarkNewsDecay(std::move(parent)) {
+        self = pybind11::reinterpret_borrow<pybind11::object>(pybind11::handle(get_object_handle(&parent, pybind11::detail::get_type_info(typeid(DarkNewsDecay)))));
+    }
+    pyDarkNewsDecay(DarkNewsDecay const & parent) : DarkNewsDecay(parent) {
+        self = pybind11::reinterpret_borrow<pybind11::object>(pybind11::handle(get_object_handle(&parent, pybind11::detail::get_type_info(typeid(DarkNewsDecay)))));
+    }
+    pybind11::object self;
+
+    double TotalDecayWidth(dataclasses::InteractionRecord const & interaction) const override {
+        SELF_OVERRIDE(
+            self,
+            DarkNewsDecay,
+            double,
+            TotalDecayWidth,
+            "TotalDecayWidth",
+            std::cref(interaction)
+        )
+    }
+
+    double TotalDecayWidthForFinalState(dataclasses::InteractionRecord const & interaction) const override {
+        SELF_OVERRIDE(
+            self,
+            DarkNewsDecay,
+            double,
+            TotalDecayWidthForFinalState,
+            "TotalDecayWidthForFinalState",
+            std::cref(interaction)
+        )
+    }
+
+    double TotalDecayWidth(siren::dataclasses::ParticleType primary) const override {
+        SELF_OVERRIDE(
+            self,
+            DarkNewsDecay,
+            double,
+            TotalDecayWidth,
+            "TotalDecayWidth",
+            primary
+        )
+    }
+
+    double DifferentialDecayWidth(dataclasses::InteractionRecord const & interaction) const override {
+        SELF_OVERRIDE(
+            self,
+            DarkNewsDecay,
+            double,
+            DifferentialDecayWidth,
+            "DifferentialDecayWidth",
+            std::cref(interaction)
+        )
+    }
+
+    void SampleRecordFromDarkNews(dataclasses::CrossSectionDistributionRecord & record, std::shared_ptr<siren::utilities::SIREN_random> random) const override {
+        SELF_OVERRIDE(
+            self,
+            DarkNewsDecay,
+            void,
+            SampleRecordFromDarkNews,
+            "SampleRecordFromDarkNews",
+            std::ref(record),
+            random
+        )
+    }
+
+    void SampleFinalState(dataclasses::CrossSectionDistributionRecord & record, std::shared_ptr<siren::utilities::SIREN_random> random) const override {
+        SELF_OVERRIDE(
+            self,
+            DarkNewsDecay,
+            void,
+            SampleFinalState,
+            "SampleFinalState",
+            std::ref(record),
+            random
+        )
+    }
+
+    std::vector<siren::dataclasses::InteractionSignature> GetPossibleSignatures() const override {
+        SELF_OVERRIDE_PURE(
+            self,
+            DarkNewsDecay,
+            std::vector<siren::dataclasses::InteractionSignature>,
+            GetPossibleSignatures,
+            "GetPossibleSignatures"
+        )
+    }
+
+    std::vector<siren::dataclasses::InteractionSignature> GetPossibleSignaturesFromParent(siren::dataclasses::ParticleType primary_type) const override {
+        SELF_OVERRIDE_PURE(
+            self,
+            DarkNewsDecay,
+            std::vector<siren::dataclasses::InteractionSignature>,
+            GetPossibleSignaturesFromParent,
+            "GetPossibleSignaturesFromParent",
+            primary_type
+        )
+    }
+
+    std::vector<std::string> DensityVariables() const override {
+        SELF_OVERRIDE_PURE(
+            self,
+            DarkNewsDecay,
+            std::vector<std::string>,
+            DensityVariables,
+            "DensityVariables"
+        )
+    }
+
+    double FinalStateProbability(dataclasses::InteractionRecord const & record) const override {
+        SELF_OVERRIDE(
+            self,
+            DarkNewsDecay,
+            double,
+            FinalStateProbability,
+            "FinalStateProbability",
+            std::cref(record)
+        )
+    }
+
+    template<typename Archive>
+    void save(Archive & archive, std::uint32_t const version) const {
+        archive(cereal::base_class<Pybind11Trampoline<DarkNewsDecay, pyDarkNewsDecay>>(this));
+    }
+
+    template<typename Archive>
+    void load(Archive & archive, std::uint32_t version) {
+        archive(cereal::base_class<Pybind11Trampoline<DarkNewsDecay, pyDarkNewsDecay>>(this));
+    }
+}; // class pyDarkNewsDecay
 
 } // namespace interactions
 } // namespace siren
@@ -77,5 +211,8 @@ public:
 CEREAL_CLASS_VERSION(siren::interactions::DarkNewsDecay, 0);
 CEREAL_REGISTER_TYPE(siren::interactions::DarkNewsDecay);
 CEREAL_REGISTER_POLYMORPHIC_RELATION(siren::interactions::Decay, siren::interactions::DarkNewsDecay);
+
+using DarkNewsDecayTrampolineType = Pybind11Trampoline<siren::interactions::DarkNewsDecay,siren::interactions::pyDarkNewsDecay>;
+RegisterTrampolineCerealMethods(siren::interactions::DarkNewsDecay, siren::interactions::pyDarkNewsDecay, DarkNewsDecayTrampolineType)
 
 #endif // SIREN_DarkNewsDecay_H
